+++ CHANGES.md
@@ -4,6 +4,7 @@
 - `/grupper` - And it's subpaths now uses clientLoaders for data fetching
 - `/arrangementer/:id` - Uses data loaders for data fetching
 - `/arrangementer/:id/registrering` - Moved to `/arragementer/registrering/:id`
+- `/wiki` - Will now redirect to the new wiki: https://wiki.tihlde.org/
 
 ### Component changes
 - Replaced old react-dropzone image upload component
@@ -11,11 +12,14 @@
   - This component change affects:
     - `/galleri`
     - `/galleri/:id`
+    - `/admin/bannere` - not tested
+    - `/profil`
+  - Need to test deletion of existing images
 
 ### Other changes
 - Replaced constate with jotai for misc hooks
 - Added ~/ import alias for src/ (required by react-router)
 - Wave svgs converted to JSX components
+- Added uploadFileWithToast and uploadFormImage utility
 
-### New Stuff Added